+++ tomcat/src/main/java/org/apache/coyote/http11/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package org.apache.coyote.http11;
 
 public enum ContentType {
     TEXT_HTML("text/html"),
