+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -12,9 +12,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.http.ContentType;
-import nextstep.jwp.http.HttpRequest;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
