+++ tomcat/src/test/java/nextstep/jwp/http/HttpRequestTest.java
@@ -3,6 +3,9 @@ package nextstep.jwp.http;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
+import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.QueryParams;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
