+++ tomcat/src/test/java/nextstep/jwp/http/QueryParamsTest.java
@@ -3,7 +3,7 @@ package nextstep.jwp.http;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import org.assertj.core.api.Assertions;
+import org.apache.coyote.http11.request.QueryParams;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
